requirements.txt
blinker==1.9.0
CacheControl==0.14.2
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
dnspython==2.7.0
email_validator==2.2.0
firebase-admin==6.6.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.67.0rc1
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
msgpack==1.1.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.1
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
rsa==4.9
setuptools==75.8.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
